+++ RELEASE-NOTES.md
@@ -1,7 +1,8 @@
 ## 0.25
 
 * #21 Implement token refresh capability for OAuth2 authentication
-* #23 Ability to register healthchecks for remote web services #23
+* #23 Ability to register healthchecks for remote web services
+* #27 OAuth2TokenAuthenticator must create its internal Client via HttpClientFactory
 
 ## 0.24